+++ src/app/components/Header/Header.scss
@@ -34,6 +34,10 @@
   }
 }
 
+.line-between {
+  border-bottom: 1px solid #edeff2;
+}
+
 .icons-right i {
   padding: 0 10px;
 }
@@ -61,8 +65,13 @@ img.logo {
   padding-top: 20px;
   justify-content: flex-start;
 
-  & > i::before {
+  .far,
+  .fa {
     color: $color-main;
+  }
+
+  .fa-phone {
+    padding-left: 10px;
   }
 
   i {
